Correct  and add new block  Users for routers with users
