+++ main.py
@@ -12,7 +12,7 @@ from src.database.db import get_db
 from src.routes.auth import router as auth_router
 from src.routes.comments import router as comment_router
 from src.routes.tags import router as tag_router
-
+from src.routes.users import router as user_router
 app = FastAPI()
 
 # Конфігурація OAuth2 для Swagger
@@ -24,7 +24,7 @@ app.include_router(auth_router, prefix='/api')
 app.include_router(photos_router, prefix='/api')
 app.include_router(comment_router, prefix='/api')
 app.include_router(tag_router, prefix='/api')
-
+app.include_router(user_router, prefix='/api')
 
 @app.get("/items/")
 async def read_items(token: Annotated[str, Depends(oauth2_scheme)]):
